# talent-match-system/docs/api/api-contract.yaml
openapi: 3.0.0
info:
  title: 人才职位智能匹配与能力图谱系统 API
  version: 1.0.0
  description: 基于AI的人才职位智能匹配与能力图谱系统API接口文档
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/resume/upload:
    post:
      summary: 上传简历
      description: 上传简历文件并创建简历记录
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume_name:
                  type: string
                  description: 简历名称
                file:
                  type: string
                  format: binary
                  description: 简历文件（PDF/DOCX）
                name:
                  type: string
                  description: 姓名
                phone:
                  type: string
                  description: 手机号
                email:
                  type: string
                  description: 邮箱
                desired_position:
                  type: string
                  description: 期望职位
                education:
                  type: string
                  description: 学历
                school:
                  type: string
                  description: 学校
                major:
                  type: string
                  description: 专业
                degree:
                  type: string
                  description: 学位
                experience:
                  type: string
                  description: 工作经验
                skills:
                  type: string
                  description: 技能（JSON数组）
                work_years:
                  type: integer
                  description: 工作年限
      responses:
        '200':
          description: 上传成功
        '400':
          description: 文件类型错误或大小超限
  /api/resume:
    get:
      summary: 获取简历列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    resume_name:
                      type: string
                    file_path:
                      type: string
                    created_at:
                      type: string
                      format: date-time
  /api/resume/all:
    get:
      summary: 获取所有简历（企业端）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /api/resume/{resume_id}:
    get:
      summary: 获取简历详情
      security:
        - bearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
        '404':
          description: 简历不存在

    put:
      summary: 更新简历
      security:
        - bearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resume_name:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                desired_position:
                  type: string
                education:
                  type: string
                school:
                  type: string
                major:
                  type: string
                degree:
                  type: string
                experience:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
                work_experience:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 更新成功
        '404':
          description: 简历不存在

    delete:
      summary: 删除简历
      security:
        - bearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '404':
          description: 简历不存在
  /api/resume/extract-resume:
    post:
      summary: 解析简历信息
      description: 通过AI解析上传的简历文件
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: 解析成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT